driver_app/requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
Babel==2.6.0
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
Django==2.2
django-extensions==2.1.6
django-formtools==2.1
django-otp==0.5.2
django-otp-yubikey==0.5.0
django-phonenumber-field==1.3.0
django-two-factor-auth==1.8.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==1.0
oauthlib==2.0.6
pbr==5.1.2
phonenumberslite==8.10.9
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pycryptodome==3.8.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyserial==3.4
python-debian==0.1.32
pytz==2019.1
pyxdg==0.25
PyYAML==3.12
qrcode==6.1
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.12.0
sqlparse==0.3.0
stevedore==1.30.0
Twisted==17.9.0
urllib3==1.22
virtualenv==16.4.0
virtualenv-clone==0.5.1
YubiOTP==0.2.2
zope.interface==4.3.2
